third_party/blink/web_tests/accessibility/readonly-expected.txt

Link Button A B C D E F G H I
1
2
 J
1
2
 K L M N
Focusable
Heading
Plain div can't be readonly
ARIA button
ARIA toggle button
ARIA link
ARIA slider
ARIA progress meter
ARIA textbox
Button

This tests which elements expose themselves as readonly. Readonly here refers to whether the item is not editable, not whether a control value can be changed vs if it's unavailable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

link1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is false


button1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is false


text1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is false


text-readonly1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is true


text-readonly2
PASS document.activeElement == element is true
PASS axElement.isReadOnly is true


checkbox1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is false


number1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is false


radio1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is false


slider1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is false


submit1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is false


combobox1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is false


listbox1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is false


textarea1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is false


textarea-readonly1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is true


textarea-readonly2
PASS document.activeElement == element is true
PASS axElement.isReadOnly is true


focusable1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is false


heading1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is false


div1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is false


aria-button1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is false


aria-togglebutton1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is false


aria-link1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is false


aria-slider1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is true


aria-progress1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is false


aria-textbox1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is true


contenteditable_root1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is false


contenteditable_button1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is false


contenteditable_root-readonly1
PASS document.activeElement == element is true
PASS axElement.isReadOnly is false


contenteditable_root-readonly2
PASS document.activeElement == element is true
PASS axElement.isReadOnly is false


contenteditable_root-readonly3
PASS document.activeElement == element is true
PASS axElement.isReadOnly is true


PASS successfullyParsed is true

TEST COMPLETE
